--- Scripts/ComponentPanelResults/DeckPanelDialogResult.cs
@@ -1,845 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Xml;
-using Godot;
-using Lizzie.AssetManagement;
-using static System.Net.Mime.MediaTypeNames;
-
-public partial class DeckPanelDialogResult : ComponentPanelDialogResult
-{
-    private LineEdit _nameInput;
-    private LineEdit _heightInput;
-    private LineEdit _widthInput;
-    private LineEdit _thicknessInput;
-
-    private HBoxContainer _customBackRow;
-
-    private ColorPickerButton _quickBackgroundColor;
-    private ColorPickerButton _quickTextColor;
-    private LineEdit _quickText;
-
-    //quick method back of token
-    private ColorPickerButton _quickBackgroundColor2;
-    private ColorPickerButton _quickTextColor2;
-    private LineEdit _quickText2;
-
-    private CheckBox _quickBackCheckbox;
-    private CheckBox _customBackCheckbox;
-
-    //private OptionButton _shapePicker;
-
-    private TabContainer _tabs;
-
-    private TextureRect _clipRect;
-    private TextureRect _textureRect;
-    private Label _label;
-
-    private const int MaxQuickSuitCount = 8;
-    private ColorPickerButton[] _quickSuitColors = new ColorPickerButton[MaxQuickSuitCount];
-    private LineEdit[] _quickSuitValues = new LineEdit[MaxQuickSuitCount];
-    private HBoxContainer[] _quickSuitRows = new HBoxContainer[MaxQuickSuitCount];
-    private OptionButton _quickSuitCount;
-
-    private ColorPickerButton _quickBackColor;
-    private LineEdit _quickBackText;
-
-    private OptionButton _cardSizes;
-
-    //Grid Tab elements
-    private ImageSelector _gridFrontImageSelector;
-    private ImageSelector _gridBackImageSelector;
-
-    private LineEdit _gridRowCount;
-    private LineEdit _gridColCount;
-    private LineEdit _gridCardCount;
-
-    private CheckButton _gridSingleBack;
-
-    private ComponentPreview _componentPreview;
-
-    private OptionButton _frontTemplatePicker;
-    private Button _editFrontTemplateButton;
-    private OptionButton _backTemplatePicker;
-    private Button _editBackTemplateButton;
-    private OptionButton _datasetPicker;
-    private Button _datasetEditorButton;
-
-    // Called when the node enters the scene tree for the first time.
-    public override void _Ready()
-    {
-        InitializeBinding();
-        InitializeStandardSizes();
-        InitializeTemplates();
-
-        ComponentType = VisualComponentBase.VisualComponentType.Deck;
-
-        HeightWidthChange(string.Empty); //just to start
-
-        QuickSuitCountChanged(_quickSuitCount.Selected);
-        GenerateQuickCards();
-        ChangePreviewCard(0);
-
-        //register for events
-        EventBus.Instance.Subscribe<TemplateChangedEvent>(TemplateChanged);
-        EventBus.Instance.Subscribe<DataSetChangedEvent>(DataSetChanged);
-        EventBus.Instance.Subscribe<ImageChangedEvent>(ImageChanged);
-    }
-
-    private void ImageChanged(ImageChangedEvent obj)
-    {
-        //reload image selectors
-        var f = _gridFrontImageSelector.SelectedImage;
-        var b = _gridBackImageSelector.SelectedImage;
-
-        _gridFrontImageSelector.SetProject(ProjectService.Instance.CurrentProject);
-        _gridBackImageSelector.SetProject(ProjectService.Instance.CurrentProject);
-
-        //try to re-select the items
-        _gridFrontImageSelector.SelectedImage = f;
-        _gridBackImageSelector.SelectedImage = b;
-
-        UpdatePreview();
-    }
-
-    private void TemplateChanged(TemplateChangedEvent obj)
-    {
-        UpdatePreview();
-    }
-
-    private void DataSetChanged(DataSetChangedEvent obj)
-    {
-        UpdatePreview();
-    }
-
-    public override void Activate()
-    {
-        var comp = GetPreviewComponent();
-        _componentPreview.SetComponent(comp, new Vector3(Mathf.DegToRad(90), 0, 0));
-        UpdatePreview();
-    }
-
-    private VcToken GetPreviewComponent()
-    {
-        string shape = "VcToken.tscn";
-
-        var scene = GD.Load<PackedScene>($"res://Scenes/VisualComponents/{shape}");
-        var vc = scene.Instantiate<VcToken>();
-
-        vc.Ready += UpdatePreview;
-        return vc;
-    }
-
-    public override void Deactivate()
-    {
-        _componentPreview.ClearComponent();
-    }
-
-    private Dictionary<string, (float W, float H, float T)> _standardSizes;
-
-    private void InitializeStandardSizes()
-    {
-        _standardSizes = new();
-        _standardSizes.Add("Poker", (63.5f, 88.9f, 0.3f));
-        _standardSizes.Add("Bridge", (57.15f, 88.9f, 0.3f));
-        _standardSizes.Add("Mini Euro", (44.45f, 63.5f, 0.3f));
-        _standardSizes.Add("Tarot", (69.85f, 120.65f, 0.3f));
-        _standardSizes.Add("Custom", (0, 0, 0));
-
-        _cardSizes.Clear();
-        foreach (var kv in _standardSizes)
-        {
-            _cardSizes.AddItem(kv.Key);
-        }
-    }
-
-    private void InitializeTemplates()
-    {
-        _frontTemplatePicker = GetNode<OptionButton>("%FrontTemplateList");
-        _frontTemplatePicker.ItemSelected += OnFrontTemplateChanged;
-        _editFrontTemplateButton = GetNode<Button>("%EditFrontTemplateButton");
-        _editFrontTemplateButton.Pressed += EditFrontTemplate;
-
-        _backTemplatePicker = GetNode<OptionButton>("%BackTemplateList");
-        _backTemplatePicker.ItemSelected += OnBackTemplateChanged;
-        _editBackTemplateButton = GetNode<Button>("%EditBackTemplateButton");
-        _editBackTemplateButton.Pressed += EditBackTemplate;
-
-        _datasetPicker = GetNode<OptionButton>("%DatasetList");
-        _datasetPicker.ItemSelected += OnDatasetChanged;
-
-        _datasetEditorButton = GetNode<Button>("%EditDatasetButton");
-        _datasetEditorButton.Pressed += EditDataset;
-
-        UpdateTemplateTab();
-    }
-
-    private void EditFrontTemplate()
-    {
-        EventBus.Instance.Publish(new ShowTemplateEditor { TemplateName = _frontTemplate?.Name });
-    }
-
-    private void EditBackTemplate()
-    {
-        EventBus.Instance.Publish(new ShowTemplateEditor { TemplateName = _backTemplate?.Name });
-    }
-
-    private void EditDataset()
-    {
-        EventBus.Instance.Publish(
-            new ShowDatasetEditor { DatasetName = _textureContext.DataSet?.Name }
-        );
-    }
-
-    private void StandardSizeChanged(long index)
-    {
-        var cardType = _cardSizes.Text;
-
-        if (!_standardSizes.TryGetValue(cardType, out var size))
-            return;
-
-        if (size.W == 0 || size.H == 0)
-            return;
-
-        _widthInput.Text = size.W.ToString("f1");
-        _heightInput.Text = size.H.ToString("f1");
-        _thicknessInput.Text = size.T.ToString("f1");
-
-        HeightWidthChange(string.Empty);
-    }
-
-    private bool _isBound;
-
-    private void InitializeBinding()
-    {
-        if (_isBound)
-            return;
-
-        _isBound = true;
-
-        _nameInput = GetNode<LineEdit>("%ItemName");
-
-        _heightInput = GetNode<LineEdit>("%Height");
-        _heightInput.TextChanged += HeightWidthChange;
-
-        _widthInput = GetNode<LineEdit>("%Width");
-        _widthInput.TextChanged += HeightWidthChange;
-
-        _thicknessInput = GetNode<LineEdit>("%Thickness");
-
-        _componentPreview = GetNode<ComponentPreview>("%ComponentPreview");
-        _componentPreview.MultiItemMode = true;
-        _componentPreview.ItemSelected += ComponentPreviewOnItemSelected;
-
-        _tabs = GetNode<TabContainer>("%TabContainer");
-        _tabs.TabChanged += t => UpdatePreview();
-
-        _cardSizes = GetNode<OptionButton>("%StandardSize");
-        _cardSizes.ItemSelected += StandardSizeChanged;
-
-        InitializeQuickBindings();
-
-        InitializeGridBindings();
-    }
-
-    private void InitializeQuickBindings()
-    {
-        //Quick suit selections - There's a better way to do this, (instantiating the lines), but for now...
-        for (int i = 0; i < MaxQuickSuitCount; i++)
-        {
-            _quickSuitColors[i] = GetNode<ColorPickerButton>($"%QuickSuit{i + 1}Color");
-            _quickSuitValues[i] = GetNode<LineEdit>($"%QuickSuit{i + 1}Contents");
-            _quickSuitRows[i] = GetNode<HBoxContainer>($"%QSRow{i + 1}");
-
-            _quickSuitColors[i].ColorChanged += color => GenerateQuickCards();
-            _quickSuitValues[i].TextChanged += t => GenerateQuickCards();
-        }
-
-        _quickSuitCount = GetNode<OptionButton>("%QuickSuitCount");
-        _quickSuitCount.ItemSelected += QuickSuitCountChanged;
-
-        _quickBackColor = GetNode<ColorPickerButton>("%QuickBackColor");
-        _quickBackColor.ColorChanged += c => GenerateQuickCards();
-
-        _quickBackText = GetNode<LineEdit>("%QuickBackText");
-        _quickBackText.TextChanged += t => GenerateQuickCards();
-    }
-
-    private void InitializeGridBindings()
-    {
-        _gridFrontImageSelector = GetNode<ImageSelector>("%FrontImageSelector");
-        _gridFrontImageSelector.ImageSelected += FrontImageSelected;
-        _gridFrontImageSelector.SetProject(ProjectService.Instance.CurrentProject);
-
-        _gridBackImageSelector = GetNode<ImageSelector>("%BackImageSelector");
-        _gridBackImageSelector.ImageSelected += BackImageSelected;
-        _gridBackImageSelector.SetProject(ProjectService.Instance.CurrentProject);
-
-        _gridRowCount = GetNode<LineEdit>("%GridRows");
-        _gridRowCount.TextChanged += t => GenerateGridCards();
-        _gridColCount = GetNode<LineEdit>("%GridCols");
-        _gridColCount.TextChanged += t => GenerateGridCards();
-        _gridCardCount = GetNode<LineEdit>("%GridCardCount");
-        _gridCardCount.TextChanged += t => GenerateGridCards();
-
-        _gridSingleBack = GetNode<CheckButton>("%GridSingleBack");
-        _gridSingleBack.Pressed += GenerateGridCards;
-    }
-
-    private string _frontGridImage;
-    private string _backGridImage;
-
-    private async void FrontImageSelected(object sender, SelectedEventArgs<Asset> e)
-    {
-        if (e.SelectedItem == null)
-        {
-            _frontGridImage = string.Empty;
-            /*
-            _frontMasterSprite = new ImageTexture(); //maybe set to blank white?
-            UpdatePreview();
-            return;
-            */
-        }
-
-        {
-            var a = e.SelectedItem;
-            _frontGridImage = a.AssetId.ToString();
-        }
-
-        //ProjectService.Instance.FetchImageAsync(a, UpdateFrontGridTexture);
-        UpdatePreview();
-    }
-
-    private async void BackImageSelected(object sender, SelectedEventArgs<Asset> e)
-    {
-        if (e.SelectedItem == null)
-        {
-            _backGridImage = string.Empty;
-            /*
-            _backMasterSprite = new ImageTexture(); //maybe set to blank white?
-            UpdatePreview();
-            return;
-            */
-        }
-        else
-        {
-            var a = e.SelectedItem;
-            _backGridImage = a.AssetId.ToString();
-        }
-        UpdatePreview();
-
-        //ProjectService.Instance.FetchImageAsync(a, UpdateBackGridTexture);
-    }
-
-    private void ComponentPreviewOnItemSelected(object sender, ItemSelectedEventArgs e)
-    {
-        _curCard = e.Index;
-        UpdatePreview();
-    }
-
-    private int _curCard = 0;
-    private List<QuickCardData> _quickCards = new();
-    private List<QuickCardData> _quickSuits = new();
-
-    private void ChangePreviewCard(int direction)
-    {
-        if (_componentPreview.ItemCount == 0)
-            return;
-
-        _curCard += direction;
-        _curCard = Mathf.Clamp(_curCard, 0, _componentPreview.ItemCount - 1);
-
-        UpdatePreview();
-    }
-
-    private void ShowPreviewCard(int cardId)
-    {
-        if (_componentPreview.ItemCount == 0)
-            return;
-
-        _curCard = cardId;
-        _curCard = Mathf.Clamp(_curCard, 0, _componentPreview.ItemCount - 1);
-
-        UpdatePreview();
-    }
-
-    private void HeightWidthChange(string newtext)
-    {
-        UpdatePreview();
-    }
-
-    private int _gridRows;
-    private int _gridCols;
-    private int _gridCount;
-
-    private void GenerateGridCards()
-    {
-        int.TryParse(_gridRowCount.Text, out _gridRows);
-        int.TryParse(_gridColCount.Text, out _gridCols);
-        int.TryParse(_gridCardCount.Text, out _gridCount);
-
-        _componentPreview.ItemCount = _gridCount;
-
-        ChangePreviewCard(0);
-    }
-
-    private int _suitCount;
-
-    private void GenerateQuickCards()
-    {
-        _suitCount = _quickSuitCount.Selected + 1;
-        LoadQuickSuits();
-
-        _quickCards = TemplateEngine.GenerateQuickCards(_quickSuits, _suitCount);
-
-        _quickCards.Clear();
-
-        for (int i = 0; i < _suitCount; i++)
-        {
-            var values = Utility.ParseValueRanges(_quickSuitValues[i].Text);
-
-            foreach (var v in values)
-            {
-                var c = new QuickCardData
-                {
-                    BackgroundColor = _quickSuitColors[i].Color,
-                    Caption = v,
-                    CardBackColor = _quickBackColor.Color,
-                    CardBackValue = _quickBackText.Text,
-                };
-
-                _quickCards.Add(c);
-            }
-        }
-
-        ChangePreviewCard(0);
-        _componentPreview.ItemCount = _quickCards.Count;
-    }
-
-    public override List<string> Validity()
-    {
-        return new List<string>();
-    }
-
-    public override Dictionary<string, object> GetParams()
-    {
-        var d = new Dictionary<string, object>();
-
-        d.Add("ComponentName", _nameInput.Text);
-        d.Add("Height", ParamToFloat(_heightInput.Text));
-        d.Add("Width", ParamToFloat(_widthInput.Text));
-        d.Add("Thickness", ParamToFloat(_thicknessInput.Text));
-        d.Add("Shape", 0);
-
-        switch (_tabs.CurrentTab)
-        {
-            case 0:
-                AddQuickParameters(d);
-                break;
-
-            case 1: //Grid
-                AddGridParameters(d);
-                break;
-
-            case 2:
-                AddTemplateParameters(d);
-                break;
-        }
-
-        return d;
-    }
-
-    private void AddQuickParameters(Dictionary<string, object> d)
-    {
-        LoadQuickSuits();
-        d.Add("QuickCardData", _quickSuits);
-        d.Add("DifferentBack", true);
-        d.Add("Mode", VcToken.TokenBuildMode.QuickDeck);
-    }
-
-    private void AddGridParameters(Dictionary<string, object> d)
-    {
-        d.Add("FrontGridImageKey", _frontGridImage);
-        d.Add("BackGridImageKey", _backGridImage);
-        d.Add("GridRows", _gridRows);
-        d.Add("GridCols", _gridCols);
-        d.Add("GridCount", _gridCount);
-
-        d.Add("Mode", VcToken.TokenBuildMode.Grid);
-        d.Add("DifferentBack", true);
-        d.Add("GridSingleBack", _gridSingleBack.ButtonPressed);
-    }
-
-    private void AddTemplateParameters(Dictionary<string, object> d)
-    {
-        d.Add("Mode", VcToken.TokenBuildMode.Template);
-
-        if (_frontTemplate != null)
-        {
-            d.Add("FrontTemplate", _frontTemplate.Name);
-        }
-
-        if (_backTemplate != null)
-        {
-            d.Add("BackTemplate", _backTemplate.Name);
-        }
-        else if (_frontTemplate != null)
-        {
-            d.Add("BackTemplate", _frontTemplate.Name);
-        }
-
-        d.Add("Dataset", _textureContext.DataSet?.Name);
-        //d.Add("CardReference", _textureContext.CurrentRowName);
-    }
-
-    private void QuickSuitCountChanged(long suitCount)
-    {
-        for (int i = 0; i < _quickSuitRows.Length; i++)
-        {
-            _quickSuitRows[i].Visible = (suitCount > i - 1);
-        }
-
-        GenerateQuickCards();
-    }
-
-    private void LoadQuickSuits()
-    {
-        _quickSuits.Clear();
-
-        _suitCount = _quickSuitCount.Selected + 1;
-
-        for (int i = 0; i < _suitCount; i++)
-        {
-            var suit = new QuickCardData
-            {
-                BackgroundColor = _quickSuitColors[i].Color,
-                Caption = _quickSuitValues[i].Text,
-                CardBackColor = _quickBackColor.Color,
-                CardBackValue = _quickBackText.Text,
-            };
-
-            _quickSuits.Add(suit);
-        }
-    }
-
-    private void UpdatePreview()
-    {
-        //normalize the size
-        var h = ParamToFloat(_heightInput.Text);
-        var w = ParamToFloat(_widthInput.Text);
-        if (h == 0 || w == 0)
-        {
-            _componentPreview.SetComponentVisibility(false);
-            return;
-        }
-
-        _componentPreview.SetComponentVisibility(true);
-
-        //normalize dimensions to 10x10x10 outer extants
-        var scale = 10f / Math.Max(h, w);
-
-        const int dpi = 250;
-
-        _textureContext.ParentSize = new Vector2(350, 250);
-        _textureContext.Dpi = 100; //TODO - Figure out what this should be
-
-        var card = _componentPreview.CurrentItem;
-
-        var d = GetParams();
-
-        Utility.UpdateParam(d, "Height", h * scale);
-        Utility.UpdateParam(d, "Width", w * scale);
-        _componentPreview.Build(d, GetRow(card), TextureFactory);
-    }
-
-    private string GetRow(int card)
-    {
-        switch (_tabs.CurrentTab)
-        {
-            case 0: //Quick
-                return (card + 1).ToString();
-
-            case 1: //Grid
-                return card.ToString();
-
-            case 2:
-                if (_textureContext.DataSet == null)
-                    return string.Empty;
-                if (card < 0 || card >= _textureContext.DataSet.Rows.Count)
-                    return string.Empty;
-                return _textureContext.DataSet.Rows.ElementAt(card).Key;
-        }
-
-        return string.Empty;
-    }
-
-    private Texture2D _frontMasterSprite;
-    private Texture2D _backMasterSprite;
-
-    private void UpdateGrid(Dictionary<string, object> d, int card)
-    {
-        d.Add("FrontMasterSprite", _frontMasterSprite);
-        d.Add("GridRows", _gridRows);
-        d.Add("GridCols", _gridCols);
-        d.Add("GridIndex", card);
-
-        d.Add("Shape", 0);
-        d.Add("Mode", VcToken.TokenBuildMode.Grid);
-        d.Add("DifferentBack", false);
-    }
-
-    private TextureContext _textureContext = new();
-
-    private void OnDatasetChanged(long index)
-    {
-        if (_datasetPicker.Selected == 0)
-        {
-            _textureContext.DataSet = null;
-            _textureContext.CurrentRowName = null;
-            _componentPreview.MultiItemMode = false;
-        }
-        else
-        {
-            _textureContext.DataSet = _currentProject.Datasets[
-                _datasetPicker.GetItemText((int)index)
-            ];
-            _componentPreview.MultiItemMode = true;
-            _componentPreview.SetItemLabels(_textureContext.DataSet.Rows.Keys.ToList());
-        }
-
-        UpdatePreview();
-    }
-
-    private Template _frontTemplate;
-    private Template _backTemplate;
-
-    private void OnFrontTemplateChanged(long index)
-    {
-        if (_frontTemplatePicker.Selected == 0)
-        {
-            _frontTemplate = null;
-        }
-        else
-        {
-            _frontTemplate = _currentProject.Templates[
-                _frontTemplatePicker.GetItemText((int)index)
-            ];
-        }
-
-        UpdatePreview();
-    }
-
-    private void OnBackTemplateChanged(long index)
-    {
-        if (_backTemplatePicker.Selected == 0)
-        {
-            _backTemplate = null;
-        }
-        else
-        {
-            _backTemplate = _currentProject.Templates[_backTemplatePicker.GetItemText((int)index)];
-        }
-
-        UpdatePreview();
-    }
-
-    private Dictionary<string, object> QuickCardParams(QuickCardData cardData)
-    {
-        var p = new Dictionary<string, object>();
-
-        p.Add("Shape", 0);
-        p.Add("Mode", VcToken.TokenBuildMode.QuickDeck);
-        p.Add("FrontBgColor", cardData.BackgroundColor);
-        p.Add("FrontCaption", cardData.Caption);
-        p.Add("FrontCaptionColor", Colors.Black);
-        p.Add("DifferentBack", true);
-        p.Add("BackBgColor", cardData.CardBackColor);
-        p.Add("BackCaption", cardData.CardBackValue);
-        p.Add("BackCaptionColor", Colors.Black);
-        p.Add("FrontFontSize", 72);
-        p.Add("BackFontSize", 24);
-
-        var fqt = new QuickTextureField
-        {
-            ForegroundColor = Colors.Black,
-            FaceType = TextureFactory.TextureObjectType.Text,
-            Caption = cardData.Caption,
-            Quantity = 1,
-        };
-        p.Add("QuickFront", fqt);
-
-        var bqt = new QuickTextureField
-        {
-            ForegroundColor = Colors.Black,
-            FaceType = TextureFactory.TextureObjectType.Text,
-            Caption = cardData.CardBackValue,
-            Quantity = 1,
-        };
-        p.Add("QuickBack", bqt);
-
-        return p;
-    }
-
-    private Project _currentProject;
-
-    public override Project CurrentProject
-    {
-        get => _currentProject;
-        set
-        {
-            _currentProject = value;
-            UpdateTemplateTab();
-        }
-    }
-
-    private void UpdateTemplateTab()
-    {
-        if (CurrentProject == null || _frontTemplatePicker == null)
-            return;
-
-        _frontTemplatePicker.Clear();
-        _backTemplatePicker.Clear();
-
-        _frontTemplatePicker.AddItem("(none)");
-        _backTemplatePicker.AddItem("(none)");
-        foreach (
-            var t in CurrentProject.Templates.Where(x =>
-                x.Value.Target == Template.TemplateTarget.Flat
-            )
-        )
-        {
-            _frontTemplatePicker.AddItem(t.Key);
-            _backTemplatePicker.AddItem(t.Key);
-        }
-
-        _datasetPicker.Clear();
-        _datasetPicker.AddItem("(none)");
-        foreach (var d in CurrentProject.Datasets)
-        {
-            _datasetPicker.AddItem(d.Key);
-        }
-    }
-
-    public override void DisplayPrototype(Guid prototypeId)
-    {
-        var prototype = ProjectService.Instance.CurrentProject.Prototypes[prototypeId];
-        DisplayPrototype(prototype);
-    }
-
-    public override void DisplayPrototype(Prototype prototype)
-    {
-        _nameInput.Text = prototype.Name;
-        _heightInput.Text = prototype.Parameters["Height"].ToString();
-        _widthInput.Text = prototype.Parameters["Width"].ToString();
-        if (prototype.Parameters.ContainsKey("Thickness"))
-            _thicknessInput.Text = prototype.Parameters["Thickness"].ToString();
-
-        if (
-            prototype.Parameters.ContainsKey("Mode")
-            && prototype.Parameters["Mode"] is VcToken.TokenBuildMode mode
-        )
-        {
-            _tabs.CurrentTab = mode switch
-            {
-                VcToken.TokenBuildMode.QuickDeck => 0,
-                VcToken.TokenBuildMode.Grid => 1,
-                VcToken.TokenBuildMode.Template => 2,
-                _ => 0,
-            };
-        }
-
-        // Handle Quick mode parameters
-        if (_tabs.CurrentTab == 0 && prototype.Parameters.ContainsKey("QuickCardData"))
-        {
-            if (prototype.Parameters["QuickCardData"] is List<QuickCardData> quickData)
-            {
-                _quickSuitCount.Select(quickData.Count - 1);
-                for (int i = 0; i < quickData.Count && i < MaxQuickSuitCount; i++)
-                {
-                    _quickSuitColors[i].Color = quickData[i].BackgroundColor;
-                    _quickSuitValues[i].Text = quickData[i].Caption;
-                }
-
-                if (quickData.Count > 0)
-                {
-                    _quickBackColor.Color = quickData[0].CardBackColor;
-                    _quickBackText.Text = quickData[0].CardBackValue;
-                }
-            }
-        }
-
-        // Handle Grid mode parameters
-        if (_tabs.CurrentTab == 1)
-        {
-            if (prototype.Parameters.ContainsKey("GridRows"))
-            {
-                _gridRowCount.Text = prototype.Parameters["GridRows"].ToString();
-            }
-
-            if (prototype.Parameters.ContainsKey("GridCols"))
-            {
-                _gridColCount.Text = prototype.Parameters["GridCols"].ToString();
-            }
-
-            if (prototype.Parameters.ContainsKey("GridCount"))
-            {
-                _gridCardCount.Text = prototype.Parameters["GridCount"].ToString();
-            }
-        }
-
-        // Handle Template mode parameters
-        if (_tabs.CurrentTab == 2)
-        {
-            if (prototype.Parameters.ContainsKey("FrontTemplate"))
-            {
-                string frontTemplateName = prototype.Parameters["FrontTemplate"].ToString();
-                for (int i = 0; i < _frontTemplatePicker.ItemCount; i++)
-                {
-                    if (_frontTemplatePicker.GetItemText(i) == frontTemplateName)
-                    {
-                        _frontTemplatePicker.Select(i);
-                        break;
-                    }
-                }
-            }
-
-            if (prototype.Parameters.ContainsKey("BackTemplate"))
-            {
-                string backTemplateName = prototype.Parameters["BackTemplate"].ToString();
-                for (int i = 0; i < _backTemplatePicker.ItemCount; i++)
-                {
-                    if (_backTemplatePicker.GetItemText(i) == backTemplateName)
-                    {
-                        _backTemplatePicker.Select(i);
-                        break;
-                    }
-                }
-            }
-
-            if (prototype.Parameters.ContainsKey("Dataset"))
-            {
-                string datasetName = prototype.Parameters["Dataset"].ToString();
-                for (int i = 0; i < _datasetPicker.ItemCount; i++)
-                {
-                    if (_datasetPicker.GetItemText(i) == datasetName)
-                    {
-                        _datasetPicker.Select(i);
-                        break;
-                    }
-                }
-            }
-        }
-
-        UpdatePreview();
-    }
-}
-
-public class QuickCardData
-{
-    public Color BackgroundColor { get; set; }
-    public string Caption { get; set; }
-
-    public Color CardBackColor { get; set; }
-    public string CardBackValue { get; set; }
-}
