+++ Scripts/VisualComponents/QuickCardData.cs
@@ -1,0 +1,10 @@
+using Godot;
+
+public class QuickCardData
+{
+    public Color BackgroundColor { get; set; }
+    public string Caption { get; set; }
+
+    public Color CardBackColor { get; set; }
+    public string CardBackValue { get; set; }
+}
